pyproject.toml
[project]
name = "livekit-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit>=1.0.25",
    "livekit-agents>=1.3.11",
    "livekit-plugins-noise-cancellation>=0.2.5",
    "livekit-plugins-silero>=1.3.11",
    "livekit-plugins-turn-detector>=1.3.11",
    "python-dotenv>=1.2.1",
]
